Impresión de niveles infinitos prematura
Parece que funciona correctamente la escala pero no hay desplazamiento en el eje x.
Se crearon fuciones de trasladar y clonar nivel

+++ nave.c
@@ -33,11 +33,11 @@ nave_t *nave_crear(void){
     return nave;
 }
 
-double nave_get_pos_x(nave_t *nave){
+float nave_get_pos_x(nave_t *nave){
     return nave->pos[X];
 }
 
-double nave_get_pos_y(nave_t *nave){
+float nave_get_pos_y(nave_t *nave){
     return nave->pos[Y];
 }
